refactor: 🏄🏽 move metrics tests to test module
