+++ tests/api/main.rs
@@ -2,4 +2,5 @@
 
 mod helpers;
 mod model;
+mod observability;
 mod startup;
